# Optimize further for large NNC case