--- a/src/opm/parser/eclipse/Parser/ParserRecord.cpp
+++ b/src/opm/parser/eclipse/Parser/ParserRecord.cpp
@@ -134,7 +134,7 @@ namespace {
             parseContext.handleError(ParseContext::PARSE_EXTRA_DATA , msg_format, location, errors);
         }
 
-        return { std::move( items ) };
+        return { std::move( items ), false };
     }
 
     bool ParserRecord::equal(const ParserRecord& other) const {