These optimizations should help in the general case, but in particular with cases are constructed using large numbers of NNCs (i.e. general graph solves not using a grid geometry).

Tested on a case with ~1.1M cells. The first commit saves 12 seconds in the deck processing. The second also saves 12 seconds in startup, and also 12 seconds in destructors at the end (it still takes 25 seconds, down from 37 seconds, to destroy the Deck object, this may deserve a separate look). @joakim-hove if there is a reason I do not see for the extra reserve, please inform. The second commit also reduces peak memory usage from approx. 43 GB to approx. 27 GB (not measured quite as precisely as the timing profile though), which could mean the difference between running and not.